pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "django>=5.2.4",
    "django-restframework>=0.0.1",
    "djangorestframework-simplejwt>=5.4.0",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "drf-yasg>=1.21.10",
    "isort>=6.0.1",
    "postgres>=4.0",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]
prod = [
    "ruff>=0.12.3",
]
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
